+++ styles/emailSignatureForm.ts
@@ -1,0 +1,10 @@
+import { TextValidator } from 'react-material-ui-form-validator'
+import { styled } from '@mui/material/styles'
+
+const StyledTextValidator = styled(TextValidator)({
+  width: 280
+})
+
+export {
+  StyledTextValidator
+}
